1:1문의 리스트

+++ src/main/java/com/ezen/sb/model/QnaModel.java
@@ -4,13 +4,13 @@ import lombok.Data;
 
 @Data
 public class QnaModel {
-	private Integer qnaNum;
+	private long qnaNum;
 	private String qnaDate;
 	private String qnaStatus;
 	private String qnaTitle;
 	private String qnaContent;
 	private String qnaFile;
 	private String qnaAnswer;
-	private Integer userNum;
+	private long userNum;
 
 }
